avant biarritz

+++ controllers/reviewController.js
@@ -1,22 +1,8 @@
 const Review = require('./../models/reviewModel');
-const catchAsynchronousError = require('../utils/catchAsynchronousError');
 const isIDValid = require('./../utils/isIDValid');
 const factory = require('./handlerFactory');
-
-exports.isIDValid = isIDValid(Review);
-
-exports.getAllReviews = catchAsynchronousError(async (req, res, next) => {
-  let filter = {};
-  if (req.params.spaceID) filter = { space: req.params.spaceID };
-  const reviews = await Review.find(filter);
-  res.status(200).json({
-    status: 'success',
-    results: reviews.length,
-    data: {
-      reviews,
-    },
-  });
-});
+const catchAsynchronousError = require('../utils/catchAsynchronousError');
+const AppError = require('./../utils/AppError');
 
 exports.setIDs = (req, res, next) => {
   if (!req.body.space) req.body.space = req.params.spaceID;
@@ -24,7 +10,18 @@ exports.setIDs = (req, res, next) => {
   next();
 };
 
-exports.createReview = factory.create(Review);
+exports.checkUserPermission = catchAsynchronousError(async (req, res, next) => {
+  if (req.user.role === 'admin') return next();
 
+  const review = await Review.findById(req.params.id);
+  if (String(review.user._id) !== String(req.user.id))
+    return next(new AppError('This is not your review', 400));
+  next();
+});
+
+exports.getAllReview = factory.getAll(Review);
+exports.createReview = factory.create(Review);
 exports.deleteOneReview = factory.deleteOne(Review);
 exports.updateOneReview = factory.updateOne(Review);
+exports.getOneReview = factory.getOne(Review);
+exports.isIDValid = isIDValid(Review);
